packages/rag/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18
PyMuPDF==1.25.1
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.14
langchain-chroma==0.2.0
chromadb==0.6.3
google-generativeai==0.8.3
python-dotenv==1.0.1
httpx==0.28.1
